Rename package to printnode-community
