+++ tests/test_imports.py
@@ -1,14 +1,14 @@
 def test_gateway_imports_from_single_namespace():
-    from printnode_api import Gateway
-    from printnode_api.gateway import Gateway as GatewayFromModule
+    from printnode_community import Gateway
+    from printnode_community.gateway import Gateway as GatewayFromModule
 
     assert Gateway is GatewayFromModule
 
 
 def test_submodules_export_expected_objects():
-    from printnode_api.auth import Unauthorized
-    from printnode_api.model import ModelFactory
-    from printnode_api.util import camel_to_underscore
+    from printnode_community.auth import Unauthorized
+    from printnode_community.model import ModelFactory
+    from printnode_community.util import camel_to_underscore
 
     assert Unauthorized.__name__ == 'Unauthorized'
     assert ModelFactory.__name__ == 'ModelFactory'
